requirements.txt
langchain
python-dotenv
crewai
poetry
